HID: logitech-dj: support sharing struct dj_receiver_dev between USB-interfaces
dj/HID++ receivers are really a single logical entity, but for BIOS/Windows
compatibility they have multiple USB interfaces. For the upcoming
non-unifying receiver support, we need to listen for events from / bind to
all USB-interfaces of the receiver.

This commit add support to the logitech-dj code for creating a single
dj_receiver_dev struct for all interfaces belonging to a single
USB-device / receiver, in preparation for adding non-unifying receiver
support.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-Do not call hid_err(djrcv_dev->hidpp, ... in delayedwork_callback in a
 case where djrcv_dev->hidpp may be NULL, instead use pr_warn in this
 special case. The move from err -> warn is intentional since this may
 happen under normal circumstances (input received during enumeration)