HID: logitech-dj: add support for 27 MHz receivers
Most Logitech wireless keyboard and mice using the 27 MHz are hidpp10
devices, add support to logitech-dj for their receivers.

Doing so leads to 2 improvements:

1) All these devices share the same USB product-id for their receiver,
making it impossible to properly map some special keys / buttons
which differ from device to device. Adding support to logitech-dj to
see these as hidpp10 devices allows us to get the actual device-id
from the keyboard / mouse.

2) It enables battery-monitoring of these devices

Note this commits also sets up a better name for 27 MHz hidpp devices,
this is done because unlike other hidpp devices, the logitech-hidpp code
will not override the name, so having a good name is important.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-Remove chunk to pick a better name for hidpp devices, this is replaced
 with a more generic fix in a separate patch