power: supply: axp288_fuel_gauge: Add ACEPC T8 mini PC to the blacklist
The ACEPC T8 Cherry Trail Z8350 mini PC uses an AXP288 and as a PC,
rather then a portable, it does not have a battery. Still for some
reason the AXP288 not only thinks there is a battery, it actually
thinks it is discharging while the PC is running, slowly going to
0% full, causing userspace to shutdown the system due to the battery
being critically low after a while.

This commit adds the ACEPC T8 to the axp288 fuel-gauge driver blacklist,
so that we stop reporting bogus battery readings on this device.

BugLink: https://bugzilla.redhat.com/show_bug.cgi?id=1690852
Cc: stable@vger.kernel.org
Signed-off-by: Hans de Goede <hdegoede@redhat.com>