HID: logitech-hidpp: add support for HID++ 1.0 mice hwheel and extra buttons
Some HID++ 1.0 (27 MHz) mice have a wheel which clicks left / right
for horizontal scrolling and some extra buttons which are not reported
through the regular mouse input report.

This commits adds support for using the special HID++ input reports for
the hwheel and extra buttons, so that these will work under Linux.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
---
Changes in v2:
-break "if (hidpp->quirks & HIDPP_QUIRK_HIDPP_HWHEEL_AND_EXTRA_BTNS)" checks
 out of the if (quirks & HIDPP_QUIRK_CLASS_FOO) ...
 else if (quirks & HIDPP_QUIRK_CLASS_BAR) ... else if ...  chain, this
 quirk is intened to possibly be combined with a device specific class,
 so it should not be an else if, but a stand-alone if