ARM: Stop using deprecated BTN_EXTRA, _SIDE, _FORWARD and _BACK defines
Because the name of the BTN_EXTRA, _SIDE, _FORWARD and _BACK defines
was a mismatch with how their values were actually being used, they
have been deprecated.

This commit moves all users of these defines under arch/arm over to
the new defines with the _same_ value.

This commit does not cause any functional changes, after this userspace
will see the exact same input_event codes as before.

Note that in case of the "user" buttons in the dts files, the original
usage of BTN_EXTRA is incorrect, as stated in
Documentation/input/event-codes.rst:

"BTN_{0,1,2,etc} are good generic codes for unlabeled buttons. Do not use
meaningful buttons, like BTN_FORWARD.", fixing this falls outside of the
scope of this commit as that would actually cause a functional change.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>