HID: Stop using deprecated BTN_EXTRA, _SIDE, _FORWARD and _BACK defines
Because the name of the BTN_EXTRA, _SIDE, _FORWARD and _BACK defines
was a mismatch with how their values were actually being used, they
have been deprecated.

This commit moves all users of these defines under drivers/hid over to
the new defines with the _same_ value.

This commit does not cause any functional changes, after this userspace
will see the exact same input_event codes as before.

Note that at least in the case of the m560 code in hid-logitech-hidpp.c
it seems that accidentally the old, not interpreted by userspace as such,
_FORWARD and _BACK defines were used. At least with Xorg this may have
worked somewhat if the BTN_EXTRA and BTN_SIDE buttons were not generated
as Xorg will remove holes from the range, so the buttons would still
be mapped to Xorg button 8 and 9, but in this case forward and back will
have been swapped. Fixing this falls outside of the scope of this commit
as that would actually cause a functional change.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>