Bluetooth: hidp: Stop using deprecated BTN_EXTRA, _SIDE defines
Because the name of the BTN_EXTRA and BTN_SIDE defines was a mismatch with
how their values were actually being used, they have been deprecated.

This commit moves the bluetooth-hidp code over to the new defines with the
_same_ value.

This commit does not cause any functional changes, after this userspace
will see the exact same input_event codes as before.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>